:sparkles: 身份认证

+++ src/server/user/user.controller.ts
@@ -1,5 +1,6 @@
-import { Controller, Get, Post, Req, Query, Body, UsePipes } from '@nestjs/common';
+import { Controller, Get, Post, Req, Query, Body, UseGuards } from '@nestjs/common';
 import { ApiTags, ApiQuery, ApiBody, ApiParam, ApiHeader, ApiHeaders, ApiResponse } from '@nestjs/swagger';
+import { AuthGuard } from '@nestjs/passport';
 import { UserService } from './user.service';
 import { User } from './entities/user.entity';
 import { CreateUserDto } from './dto/create.user.dto';
@@ -7,6 +8,8 @@ import { UpdateUserDto } from './dto/update.user.dto';
 import { RegisterUserDto } from './dto/register.user.dto';
 import { LoginUserDto } from './dto/login.user.dto';
 import { DeleteUserDto } from './dto/delete.user.dto';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
+import { LocalAuthGuard } from '../auth/guards/local-auth.guard';
 
 @ApiTags('用户')
 @Controller('user')
@@ -14,29 +17,24 @@ export class UserController {
   constructor(private readonly userService: UserService) {
   }
 
-  @Post('login')
-  @ApiResponse({ status: 200, description: '请求成功！' })
-  login(@Body() loginUserDto: LoginUserDto): Promise<CreateUserDto> {
-    return this.userService.login(loginUserDto);
-  }
-
   @Post('register')
-  @ApiResponse({ status: 200, description: '请求成功！' })
-  register(@Body() registerUserDto: RegisterUserDto): Promise<CreateUserDto> {
-    return this.userService.register(registerUserDto);
+  async register(@Body() registerUserDto: RegisterUserDto): Promise<CreateUserDto> {
+    return await this.userService.register(registerUserDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Post('add')
-  @ApiResponse({ status: 200, description: '请求成功！' })
-  add(@Body() createUserDto: CreateUserDto): Promise<CreateUserDto> {
-    return this.userService.insert(createUserDto);
+  async add(@Body() createUserDto: CreateUserDto): Promise<CreateUserDto> {
+    return await this.userService.insert(createUserDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Get('findList')
-  findList(): Promise<User[]> {
-    return this.userService.selectList();
+  async findList(): Promise<User[]> {
+    return await this.userService.selectList();
   }
 
+  @UseGuards(JwtAuthGuard)
   @Get('findListPage')
   @ApiQuery({
     name: 'page',
@@ -48,34 +46,51 @@ export class UserController {
     description: '每页条数',
     required: false,
   })
-  findListPage(@Query('page') page: number, @Query('limit') limit: number): Promise<any> {
+  async findListPage(@Query('page') page: number, @Query('limit') limit: number): Promise<any> {
     page = page ? page : 1;
     limit = limit ? limit : 10;
-    return this.userService.selectListPage(page, limit);
+    return await this.userService.selectListPage(page, limit);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Get('findById')
-  findById(@Query('id') id: string): Promise<User> {
-    return this.userService.selectById(id);
+  async findById(@Query('id') id: string): Promise<User> {
+    return await this.userService.selectById(id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Post('modify')
-  modify(@Body() updateUserDto: UpdateUserDto): Promise<void> {
-    return this.userService.update(updateUserDto);
+  async modify(@Body() updateUserDto: UpdateUserDto): Promise<any> {
+    let { id } = updateUserDto;
+    let entity = await this.userService.selectById(id);
+    if (entity) {
+      return this.userService.update(updateUserDto);
+    } else {
+      // todo
+      return `数据 id = ${id} 不存在！`;
+    }
   }
 
+  @UseGuards(JwtAuthGuard)
   @Post('remove')
   @ApiBody({
     schema: {
       type: 'object',
       properties: {
         id: {
+          type: 'string',
           description: '主键 id',
         },
       },
     },
   })
-  remove(@Body('id') id): Promise<void> {
-    return this.userService.deleteById(id);
+  async remove(@Body('id') id: string): Promise<any> {
+    let entity = await this.userService.selectById(id);
+    if (entity) {
+      return await this.userService.deleteById(id);
+    } else {
+      // todo
+      return `数据 id = ${id} 不存在！`;
+    }
   }
 }
